FFA-78 Refactor Names
